// src/types/api.ts
export interface components {
    schemas: {
// ...
        PlatformLoginRequest: {
            /**
             * Email
             * Format: email
             */
            email: string;
            /** Password */
            password: string;
        };
        /** PlatformLogoutResponse */
        PlatformLogoutResponse: {
            /**
             * Session Id
             * Format: uuid
             */
            session_id: string;
            /** Status */
            status: string;
        };
        /** PlatformMFAEnableRequest */
        PlatformMFAEnableRequest: {
            /** Totp Code */
            totp_code: string;
        };
        /** PlatformMFAEnableResponse */
        PlatformMFAEnableResponse: {
            /** Recovery Codes */
            recovery_codes: string[];
        };
        /** PlatformMFASetupResponse */
        PlatformMFASetupResponse: {
            /** Secret */
            secret: string;
            /** Provisioning Uri */
            provisioning_uri: string;
        };
        /** PlatformMFAVerifyRequest */
        PlatformMFAVerifyRequest: {
            /** Mfa Challenge Token */
            mfa_challenge_token: string;
            /** Totp Code */
            totp_code?: string | null;
            /** Recovery Code */
            recovery_code?: string | null;
        };
// ...
        PlatformTokenResponse: {
            /** Access Token */
            access_token?: string | null;
            /** Refresh Token */
            refresh_token?: string | null;
            /**
             * Token Type
             * @default bearer
             */
            token_type: string;
            /** Expires In */
            expires_in: number;
            /** Refresh Expires In */
            refresh_expires_in?: number | null;
            /** Session Id */
            session_id?: string | null;
            /**
             * Mfa Required
             * @default false
             */
            mfa_required: boolean;
            /**
             * Mfa Enrollment Required
             * @default false
             */
            mfa_enrollment_required: boolean;
            /** Mfa Challenge Id */
            mfa_challenge_id?: string | null;
            /** Mfa Challenge Token */
            mfa_challenge_token?: string | null;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}